refactor manage samples permissions again, so that addons permissions can be added

+++ app/controllers/concerns/sample_actions.rb
@@ -11,7 +11,7 @@ module SampleActions
       params[:sample_ids].each do |id|
         sample = Sample.find_by_id(id)
 
-        if sample
+        if sample && can_update_or_delete_sample?(sample)
           sample.destroy
           counter_user += 1
         else
@@ -45,8 +45,6 @@ module SampleActions
   end
 
   def check_destroy_samples_permissions
-    unless can_manage_sample?(@project.team)
-      render_403
-    end
+    render_403 unless can_manage_samples?(@project.team)
   end
 end
